requirements.txt
einops==0.7.0
matplotlib==3.5.1
monai==1.3.0
nibabel==5.2.0
numpy==1.24.2
opencv-python==4.5.4.60
pandas==2.2.0
pycocotools==2.0.6
pynrrd==1.0.0
safetensors==0.4.1
scikit-image==0.22.0
scikit-learn==1.0.2
scipy==1.12.0
segment-anything==1.0
shellingham==1.5.4
SimpleITK==2.3.1
six==1.16.0
sklearn==1.3.2
slicerio==0.1.8
timm==0.9.12
torch==2.0.0+cu117
torchio==0.19.3
torchvision==0.15.1+cu117
tqdm==4.62.3